requirements.txt
python-telegram-bot[job-queue]
aiohttp
python-dotenv
sqlite3-api
tenacity
structlog